requirements.txt
# Azure SDK
openai==1.12.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0
azure-ai-formrecognizer==3.3.2
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.0

# LangChain
langchain==0.1.9
langchain-openai==0.0.8
langchain-community==0.0.24

# Data Processing
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1

# Visualization
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.19.0

# Web Development
fastapi==0.109.2
uvicorn[standard]==0.27.1
streamlit==1.31.1
requests==2.31.0

# Database
pyodbc==5.1.0
sqlalchemy==2.0.27

# Utilities
jupyter==1.0.0
jupyterlab==4.1.2
ipywidgets==8.1.2
tqdm==4.66.2
pyyaml==6.0.1

# Testing
pytest==8.0.2
pytest-asyncio==0.23.5

# Code Quality
black==24.2.0
flake8==7.0.0
mypy==1.8.0

